Implemented file store rollback.
As a bonus, the hyperdb is now capable of storing more than one file
per node - if a property name is supplied, the file is called
designator.property.

I decided not to migrate the existing files stored over to the new naming
scheme - the FileClass just doesn't specify the property name.