Implemented the destroy() method needed by the session database...
(and possibly others). At the same time, I removed the leading
underscores from the hyperdb methods that Really Didn't Need Them.

The journal also raises IndexError now for all situations where there is a
request for the journal of a node that doesn't have one. It used to return
[] in _some_ situations, but not all. This _may_ break code, but the tests
pass...