Date arithmetic fixes.
Date +/- Interval passes all tests again, after fixing a couple of the
tests to actually reflect the calendar used on my planet rather than
where-ever Richard was when he wrote the test <wink>

The basic problem was that when going backwards, the code was adding
the days of the current month, rather than the previous month.
There's still a bug in the testDateSubtract that I'll fix next.

Bugfix candidate (probably)