The whole filter method was replicated in back_mysql.py from rdbms_common.py
I've factored the reason for the difference, mysqls inability to do
subselects. We now have a _subselect method that is overridden in
back_mysql and most of the filter method stays in rdbms_common. This
passed all tests for mysql, postgresql and sqlite with python2.4 and
python2.3.

I've done this as a prerequisite for doing some surgery on filter.