Added two new tests for Links and Multilinks in HTMLItems:
An Item, when dereferencing a Link or Multilink should return the
correct items from the database.

Now if we name the linked items with numeric names (in the Link
example we have a status with id='1' and name='2' and vice-versa, in
the Multilink example we have the same for keyword) and the property
'name' of these items is also the key, it still works for Links but
not for Multilinks: The multilink in this case returns the keyword
with the *name* '1', not the one with the *id* '1'. I consider this a
bug and have implemented the test before looking deeper into it...