lowercase tracker name during roundup-server's cgi handling
We lowercase the tracker name when we store it in the config, we might
as well lowercase it on the incoming GET so it will match what we've
stored. This doesn't make the web interface truly case sensitive (which
would require a lot more work) but it will allow users to do something
like:

 roundup-server TaskManager=/var/tm
 wget http://server/TaskManager

and get something nicer than a 404.

This fixes [SF#1143803]