More SSL fixes.
SSL wants the underlying socket non-blocking. So we don't call
socket.setdefaulttimeout in case of SSL. This apparently now never
raises a WantReadError from SSL.  This also fixes a case where a
WantReadError is raised and apparently the bytes already read are
dropped (seems the WantReadError is really an error, not just an
indication to retry).