Fix bug in (Date- or Interval-) Range parsing
Obscure case with only a "from" is now parsed correctly.
Obviously it isn't possible to get the three cases with a missing "from"
*or* a missing "to" or *both* (in which case the interval has zero
length) right in a single regex. This also fixes broken unittests
testFilteringRangeInterval and testFilteringRangeTwoSyntaxes (which used
that syntax). This was introduced in 07c59221f363 but before that ranges
with only a "to" didn't work.