Rename TestMessage to ExampleMessage
When py.test is searching for tests it detects TestMessage as a
potential class with test cases because it has 'Test' in its name.
Because TestMessage contains a __init__ constructor py.test returns a
warning that it can't collect tests from it. Lets rename the class to
ExampleMessage to avoid this issue.