The patch to implement:
  Validate properties specified for sorting and grouping in index
  views. Original patch from martin.v.loewis via:
  https://hg.python.org/tracker/roundup/rev/439bd3060df2
  Applied by John Rouillard with some modification to properly
  identify if the bad property is a sort or grouping property. Tests
  added.

has an issue with the current code base. Apparently sometime it can be
entered without self.classname being defined. As a result the property
lookup fails. So guard it by checking for self.classname in a couple
of spots and if self.classname is not set just append the property
and let the target action sort it out.