issue2550864 - Potential information leakage via journal/history
Fix this by making the hyperdb::Class::history function check for view
permissions on the journaled properties. So a user that sees [hidden]
for a property in the web interface doesn;t see the property changes
in the history.

While doing this, relocated the filter for quiet properties
from the templating class to the hyperdb.

Also added the skipquiet option to the history command in
roundup-admin.py to enable filtering of quiet params.

Also changed calls to history() in the backend databases to report all
items.

Changed inline documentation for all history calls that document the
actions. The create action (before nov 6 2002) used to record all
parameters. After that point the create call uses an empty dictionary.
The filtering code depends on the create dictionary being empty.
It may not operate properly on very old roundup databases.

Changed calls to logging.getLogger to roundup.hyperdb.backends to
allow filtering the back end while keeping hyperdb logging.

In cgi/templating.py, changed history() function consolidating
handiling of link and unlink actions

Added tests for quiet property filtering and permission filtering
of history.