Avoid errors from selecting "no selection" on multilink (issue2550722).
As discussed in issue 2550722 there are various cases where selecting
"no selection" on a multilink can result in inappropriate errors from
Roundup:

* If selecting "no selection" produces a null edit (a value was set in
  the multilink in an edit with an error, then removed again, along
  with all other changes, in the next form submission), so the page is
  rendered from the form contents including the "-<id>" value for "no
  selection" for the multilink.

* If creating an item with a nonempty value for a multilink has an
  error, and the resubmission changes that multilink to "no selection"
  (and this in turn has subcases, according to whether the creation
  then succeeds or fails on the resubmission, which need fixes in
  different places in the Roundup code).

All of these cases have in common that it is expected and OK to have a
"-<id>" value for a submission for a multilink when <id> is not set in
that multilink in the database (because the original attempt to set
<id> in that multilink had an error), so the hyperdb.py logic to give
an error in that case is thus removed.  In the subcase of the second
case where the resubmission with "no selection" has an error, the
templating code tries to produce a menu entry for the "-<id>"
multilink value, which also results in an error, hence the
templating.py change to ignore such values in the list for a
multilink.