issue2551069 - when unsupported type is found report type
When reporting an unsupported response type, error 406, report what
type is seen by the code. Allows the user to better debug the
query. E.G. look at the url and see if they mistyped it (e.g. by
leaving a ? out of the url), or a mispelled accept header value or
....