Cleanups for bandit
Set bandit to ignore use of md5.

Treat schema from database as trusted to eval. We wrote it based on
the schema.py file.

Replace some bare except: with a proper exception.

mode 775 for index directory is correct. Allows indices to be written
by email and web interfaces that run as different users but can have
the roundup group in common.