Make the charset configurable for mysql
We now have an option to configure the mysql charset used when opening
the database. The default is now utf8 for both, python2 and python3.
No charset argument is passed if this is configured to 'default'.
Previously we were passing 'utf8' for python3 and no charset option for
python2 which made regression tests fail depending on the mysql
configuration.