Fix string conversion from database
The sql_to_hyperdb_value table would convert strings coming from the
database using "str". This is now changed to "us2s", converting unicode
strings coming from the database to utf-8. This happens only for
python2, for python3 the database interface uses unicode (aka str in
python3). This fixes the behaviour for mysql if utf8 encoding is chosen
explicitly, it seems in that case the database interface returns values
as python2 unicode objects.