Summary: Constrain format of classname and document
Define the format of a valid classname. Updated design doc, function
doc and glossary to document format.

Error message for case where we have a redefinition of a classname now
says why ValueError is raised since we are raising ValueError for
incorrect classname format as well.

Tests for all cases including verification of cause for ValueError
exceptions.