Initial implementation of Computed property
It supports query/display in html, rest and xml interfaces.
You can specify a cache parameter, but using it raises
NotImplementedError.

It does not support: search, sort or grouping by the computed field.

Checking in on a branch to get more eyeballs on it and maybe some
people to help.