issue2551142 - Import ... unique constraint failure.
Full title: Import of retired node with username after active node
fails with unique constraint failure.

Fix this in two ways:

1) sort export on keyname, retired status so that retired nodes for a
   given keyname are before the acive node in the export file.

This stops generating a broken export.

2) handle importing a broken export by deactivating/fixing up/clearing
the active record's unique index entry temporarily. Redo the
import of the retired node and resetting the active record to active.

The fixup changes the unique index (keyvalue, __retired__) from
(keyvalue, 0) to (keyvalue, -1). Then it retries the failed import of
a retired record with keyvalue. I use -1 in case something goes wrong,
It makes the record stand out in the database allowing hand recovery
if needed. Rather than using -1 I could just use the id of the record
like a normal retirement does.

If the retry of the import fails (raises exception), reset the active
record from -1 back to 0 and raise the exception.

If it succeeds, reset the active record from -1 back to 0 and continue
the import process.

Reset __retired__ from -1 to 0 on every import. I don't think the
performance loss from resetting on every exception matters as there
should be very few exceptions. Also this makes the code more
understandable.  There is no reason to leave the -1 value in place and
do a bulk rest of -1 to 0 after the class csv file is loaded.

Also if a fixup is needed it is logged at level info with the rest of
the database logging. Also success of the fixup is logged. Fixup
failure generates a propagated exception.