Handle configparser.InterpolationSyntaxError
Under Python 3, an option value with a single % (e.g. this % is a
test) throws

   configparser.InterpolationSyntaxError: '%' must be followed by
      '%' or '(', found: '%s))'

Added code to capture this, raise a different exception. roundup-admin
handles the error and exits cleanly. Other code shows the traceback.

The new error message reports the file, section and option causing the
problem to allow easier repair.

Also updated roundup translations and added tests.