Add error templates for 400 and label each item optional
Added documenttion on adding a _generic.400.html template which can be
used to display errors in FTS query syntax. It can include the chrome
supplied by page.html which makes navigating etc. easier.

Also labeled each section title in the current release optional as
none of them are required upgrades.

If we add a required step that will break a tracker if not done, it
should be labeled required and put at the top of the upgrading
section.