issue2551189 - increase text search maxlength
This removes I think all the magic references to 25 and 30 (varchar
size) and replaces them with references to maxlength or maxlength+5.

I am not sure why the db column is 5 characters larger than the size
of what should be the max size of a word, but I'll keep the buffer
of 5 as making it 1/5 the size of maxlength makes less sense.

Also added tests for fts search in templating which were missing.
Added postgres, mysql and sqlite native indexing backends in which to
test fts. Added fts test to native-fts as well to make sure it's
working.

I want to commit this now for CI.

Todo:

 add test cases for the use of FTS in the csv output in
 actions.py. There is no test coverage of the match case there.

 change maxlength to a higher value (50) as requested in the ticket.

 Modify existing extremewords test cases to allow words > 25 and < 51

 write code to migrate column sizes for mysql and postgresql to match
   maxlength I will roll this into the version 7 schema update that
   supports use of database fts support.