try to update to python 3.10.4 from 3.10.0 and try 3.11
Travis ci doesn't seem to tell us what python versions are
supported. The one form that is supposed to show that doesn't work for
me. Also it looks like nightly is suck at 3.10.0a5+ which isn't
nightly.

python 3.10.4 is released and current. 3.11 is pre-release/nightly so
trying it to replace broken nightly.

These upgrades may need to be reverted if invalid.