wsgi can cache tracker instance enabled by feature flag.
Patch by Marcus Priesch caches a loaded tracker instance and reuse it
for future client sessions.

It is enabled by a feature flag in wsgi.py since it arrived during the
2.2.0 beta period.

The provided wsgi.py is modified to enable it.  Testing is run with
flag enabled and disabled.

Ralf Schlatterbeck and Marcus tested it on one of their larger more
complex trackers and it sped up the response time to a client request
by a factor of 3 (270ms down to about 80-85ms).