Handle build issues, update css for mobile
I added www.roundup-tracker.org as a property of mine in google search
console. It reports bad mobile experience.

According to the goatcounter tracker I put in, we get 400 hits from
phones/small tablets; 400 from tablets/small laptops and 800 from
computer screens. So 1/4 is mobile and 1/2 is not a large computer
screen.

On mobile (<960px) the left hand menu is stacked on top of the column
in mobile sizes. Sub-menu links are spaced apart to make them easier
to clkick on with a finger. Same with the table of contents in the
documents. The main document content is now full display width (rather
than scrunched to the right side of the display). This is just a quick
hack, but it should make the docs more usable.

As I worked, I found that _static/style.css changes were not being
copied into the html build directory when sphinx was rerun. Nuke
entire html subdir and rebuild from scratch each time.  Also added
comments to Makefile.

Also robots.txt was being added as extra_html by sphinx, but it causes
a missing from TOC error that exits the build (when using -W). Since
exiting on warning is better, I changed Makefile to add
robots.txt. Removed robots.txt references from conf.py.