Improve perf: css/js to end; templating for goatcounter src
Lighthouse reports a lot of time was spent waiting for js and css to
load before page could be shown. Especially for customizing.html.

Keep style.css in head, but move pygments.css to end of body.
(note, lighthouse was reporting that contrast for pygments wasn;t as
high as it should be.)

Also move the 4 js files: jquery, underscore.js, doctools.js
language_data.js to the end of body.

Move the js script that shows the searchbox there too as it uses
jquery.

Use the templating pathto() function to generate path to goatcounter
js file.