Docker fix healthcheck; allow modules; cleanup; set uid
The docker healthcheck was hardcoded to check the /issues/ tracker.
Replace healthcheck with one that looks for the tracker names on the
roundup-server command line and checks the first one.

During build, additional modules can be specified using
--build-arg="pip_mod=requests setproctitle". This lets the user add
modules unique to the tracker without having to 'docker commit' a new
image from a running container.

Use --build-arg="roundup_uid=2000" to change the uid roundup runs
as. The default is 1000. This is done at build time, not run time.

Remove the sphinx package. All the dependent packages were removed
before, but sphinx wasn't. This led to spurious warnings fom the pip
dependency resolver.

Update docs with changes.