Script dumps dbm files without marshalled data like _ids
Allow the _ids database to be dumped. The _ids db (when using the
anydbm backend) records the number last used for each object
type. E.G. if there were 8 statuses defined, it would have the entry:

  status: 8

Allow other dbm databases (like the node or journal dbs) to be dumped.

Used argparse for command line options and added support for:

  dump all keys in sorted order
  dump specific key(s)
  pretty print output