set method doesn't include user set timestamp if update
set() is supposed to update the record with the key if it already
exists.

It does update the value marshalled data blob. However it doesn't
update the timestamp column for rdbms tables if provided.

This change updates the x_time column with the provided __timestamp or
preserves the original timestamp.