Change method for settin indexer; have test_livetest for pg cleanup
Add code to defer opening the indexer only if indexer is native-fts.
See if this fixes the sqlite OperationalError.

Also under python 2.7 (only), the db from test_livetracker when using
postgres FTS didn't empty the db. This caused the following
test_postgres.py test to fail.

Why it only showed up on 2.7 and not any of the 3.x releases is a
mystery.