Replace hardcoded python version with ARG variable
The Dockerfile had a hardcoded Python version in install paths and COPY
command in the second stage build.

Replace that with a hardcoded ARG pythonversion. Add RUN command to
verify pythonversion against the actual version installed in the
docker image.

If they do not match generate an appropriate --build-arg for the
docker build command that will match the docker image version.

Document the use of the variable and why it's needed in
installation.txt. Also fix typos referencing --build-args and bad
values for the arguments.