Remove duplicate h1 headers from website
Each page has an H1 header in the main part of the page. The class=header
block also had an H1 enclosing "Roundup". This isn't useful as an H1
header. The page title includes Roundup as well so I demoted the
class=header H1 to a label but styled it the same.