Modify code to reduce runtime when testing
The prior change to set default number of PBKDF2 rounds to 2000000
(2M) raised runtime in CI from 12 minutes to an hour.

This commit checks to see if we are invoked from a pytest test using:

    if ("pytest" in sys.modules and
        "PYTEST_CURRENT_TEST" in os.environ):

when no config object is present. I assume that the number of times we
have a full config object is less than with a missing config object.

See if this brings CI runtimes back down. It reduces runtimes on my
local box, but....

Code adapted from
https://stackoverflow.com/questions/25188119/test-if-code-is-executed-from-within-a-py-test-session/44595269#