auto update 'password_pbkdf2_default_rounds'                        "
If the current value for password_pbkdf2_default_rounds is the old
default value of 10000, updateconfig will use the new default
value when writing the new config.

If the curreny value for password_pbkdf2_default_rounds is less than
the new default the user will be told to update the value.

New translation strings added. Updates to upgrading.txt doc.