Test self.db.db_version_updated in schema upgrade paths
A user on the roundup-users mailing list reported that a
'roundup-admin migrate' reported that changes were needed. But it
should have done the migration.

roundup-admin uses the property db.db_version_updated to tell if there
was a migration. Testing never looked at that property. Add a test for
that property in all migration updates.

AFAICT the test is always successful. So no bug uncovered by these
tests.