Refeactor PBKDF2 and PBKDF2S5 to reuse code; fix python password.py
90% of the code for these two paths in encodePassword is identical.
Combine the identical parts of the code.

When password.py is called directly, it runs tests. But the sys.path
is messed up so that roundup/cgi is imported during execution rather
than pythonlib/cgi.py leading to an import error during setup.

This issue does not occur if the tests are run under pytest.