fix initialization of option_validator
When using option_validator to validate config.ini entries for
detectors or extensions, the list of validators was not reset when the
tracker was re-opened.

This led to the second command run in roundup-admin crashing when the
main config.ini file was read. On second tracker open/read
option_validator lists an extension/detector option that doesn't exist
in the main config.ini. When it tried to validate, the missing option
caused a crash.

This only happens if a validator is added in an extension/detector.
If the only validators ae ones for the main config.ini everything
works fine across multiple tracker opens.