Formatting fixes
Fix some blocks that should be code blocks but weren't.

Use smaller type for larger/wide code blocks by replacing:

  stuff::

     wide code here

with:

  stuff:

  .. code::
     :class: big-code

     wide code here

The big-code class uses a smaller font size so it fits the width of
the page without overflow.

Fix an ordered list that was indented more than it should have been.