Ad skiplink to main on moblie; fix overlapping links
On mobile, the Roundup label at the top of the stacked table of
contents becomes a link to jump to the top of the main body of the
page. Without this there is no indication that a link on the main menu
has been clicked and there is a lot of scrolling to get to the real
content of the page. Ideally the left sidebar would become a menu link
an be shown only when activated but ....

Ran the top level pages through lighthouse to see how they do on
mobile.  Style changes added to fix links that overlap; fix links that
were incorrectly displayed as blocks rather than in running text.

Fixes some spacing issues with nested lists inside definition lists
definition (dd) tags. Removed some indent when TOC follows n H1
header. Wrap search box on smaller screens so the search box doesn't
overlap the jump link. Add back left margin to footer.

Tried to prevent overlap with the prev/next/index submenu. It's better
but still not good.