Run coveralls push only if all builds succeed.
I wanted to run the coveralls push even if there was only one
successful run. So I forced it to run even if there was an error in
one of the paralle workflows. However coveralls push was succeeding
and setting the job status to ok even though I want it to be failing.

It appears I can't get what I want in an easy manner so don't upload
any data to coveralls if any CI test job fails.

I will still have the successful jobs uploaded to codecov at least.