Fix test failure caused by making genconfig trackerless
When I removed the need to have a tracker for genconfig, I broke a
couple of tests. The breakage was hidden by a bad workflow.

This change allows the tests to pass and refactors reporting of
UsageError exceptions in admin.py.

It also adds a test for failure in install command as there is a test
case for it and this test tests the refactor.