Make anydbm indexer handle missing items the same as rdbms backend
The anydbm Class::index() method was capturing the IndexError rather
than letting it bubble up. This is different from the rdbms backends.

In implementing batch/incremental indexing via roundup-admin, I wanted
to report any missing items that were specified in the batch. This
worked by capturing the IndexError. It worked with rdbms backends but
not anydbm. This change makes it work with anydbm as well.

Focused testing doesn't show any regressions, so committing the
change.