Reformat inclusion of COPYING.txt into license.txt.
Include as non-literal text. This removes the need to scroll to read
the license. Also makes the hyperlinks work.

Colorize the background under the included file to allow easy
identification of part of license.txt that is from COPYING.txt like we
used to have with the literal inclusion.

Had to add raw html to create a div to add the color. Using
.. container:: didn't work since indenting .. include:: caused sphinx
to throw errors on the headers in COPYING.txt.

Reformated COPYING.txt to eliminate unwanted indentation caused by
blockquote being triggered.