Clarify docker run commands and link to admin manual section
I had documented how to run a demo docker container and added a
footnote on how to change it to run with the hub.docker.com
container. It didn't really save any space and prevented
cut/paste. Replace with two examples that can be properly cut and
pasted.

Link to the section in the admin guide that describes how to
initialize a tracker unattended without disclosing the password on
command line or in environment.

Also fix a typo.