Speed up postgresql by disabiling durability settings.
Read:

  https://pythonspeed.com/articles/faster-db-tests/

I'm not using docker for postgresql at this time, but this article
describes how to start a docker postgres image with disabled settings
that may be of use if I end up having to change python 2.7 testing to
use dockerized containers against dockerized database backends.

However as a result of the above article, I took another look at:

  https://www.postgresql.org/docs/current/non-durability.html

and switched off a few more settings.

Also added some similar settings for mysql that was not in the
deployed workflow.

[skip travis]