Disable all but 2.7 and nightly testing.
I have to keep pestering travis to refill the credits I need for
testing.

Since github is testing the 3.x stuff, I only need to use travis for
latest devel (nigthly) and 2.7 python. 2.7 support has been dropped
from github.

This will hopefully reduce the number of times that I look to travis
for results and find that it didn't run because of exhausted cedits.