test: Modify testRestRateLimit test to report when system is too slow.
Add an explicit check on the runtime and an error message that reports
that the runtime was exceeded for the test to complete as written.

testRestRateLimit requires that it finish within 3 seconds. Otherwise
the number of remaining requests in the rate limit does not decrease
on every call. If disk I/O is high, the anydbm version of this test
can take > 3 seconds and result in a failed test.

My other alternative was to measure the runtime and adjust the test to
match the values that are returned. This seems like too much work and
is unlikely to be an issue outside of a developers under powered
system.