prevent <newline tag mangling
remove charref and entityref handlers; change where newline inserted
after start tag.

A test under python2 ended up with a newline between the opening '<'
and the tag. This was caused by a string that had escaped &gt; and
&lt;.

   embedded code block &lt;pre&gt\n\n<pre> python\nline 1\nline 2\n</pre>

The code was mapping &lt; etc back to < and > and confusing the parser
as to where the tag really started. It inserpreted the real pre tag as
data and inserted a newline.