Fix tests
The testEmailBodyUnchangedNewIsYes was using a signature (sig) that
didn't have a blank line before it. Therefore the sig wasn't
recognised as a message section and wasn't a candidate to be removed
even if EMAIL_LEAVE_BODY_UNCHANGED was set to no.

Appended a newline before the signature block so testing with

   self.instance.config.EMAIL_LEAVE_BODY_UNCHANGED = 'no'

will cause test to fail and setting it to yes (the test condition)
will cause test to pass by keeping signature.

Make same sig change to testEmailBodyUnchangedFollowupIsYes for same
reason.  Add new tests for other permutations:

  testEmailBodyUnchangedNewIsNew - keep sig
  testEmailBodyUnchangedFollowupIsNew - remove sig
  testEmailBodyUnchangedFollowupIsNo - remove sig