flake8: rename loop variable in 'for sendto in sendto:'
Flake8 reported 'B020 Found for loop that reassigns the iterable it is
iterating with each iterable value.'

Renamed loop variable to to_addr. There is a similar construct with a
loop over bcc_sendto with a 'bcc' loop variable. So I assume the loop
varaible can be chnaged w/o issue.

Codecov shows all the affected lines are being tested and the tests I
ran with testmon that should cover that code all passed.

We shall see if a full CI run passes.