refactor(api): early return if REST rate limit is exceeded
If the rate limit at the top of dispatch() is exceeded, return rather
than running to final return at end of method.

This does change a couple of things:

  1) output format is always in json
  2) json is alays pretty printed

Previously, the output format requested by the Accept header or format
extension in the URL was used for 1. Similarly value of @pretty was
used to control 2.

I am trying to reduce the complexities in this routine with the goal
of fixing: issue2551289 to return 406 error if the Accept
header/format extension is incorrect before executing the request.