build: xapian under 3.12 - missing version info in build components
The build with 1.4.18 fails to get the shared library file names
correct. Trying with 1.4.22 as the release against the 1.4.18 dev
libraries to see if it works better.

The sed script I have that munges the configure script to change the
method used for getting the components should work but it obviously
doesn't.

Failure under 3.12:

/usr/bin/install -c -m 644 xapian/__init__.py \
   ./xapian/__pycache__/__init__..pyc \
   ./xapian/__pycache__/__init__. \
   '/opt/hostedtoolcache/Python/3.12.0-rc.2/x64/lib/python3.12/site-packages/xapian'
/usr/bin/install: cannot stat './xapian/__pycache__/__init__..pyc': No such file or directory
/usr/bin/install: cannot stat './xapian/__pycache__/__init__.': No such file or directory

Success under 3.11 which also needs the sed script:

/usr/bin/install -c -m 644 xapian/__init__.py \
  xapian/__pycache__/__init__.cpython-311.pyc \
  xapian/__pycache__/__init__.cpython-311.opt-1.pyc \
 '/opt/hostedtoolcache/Python/3.11.5/x64/lib/python3.11/site-packages/xapian'