build: try 3.12 xapian build again.
Well configure needs to be taught not to use imp library under 3.12
as imp doesn't exist anymore.

Update sed script for configure to make the required modifications.

This should work going back to 3.6 as the replacement:

  imp.gettag() -> sys.implementation.cache_tag
  imp.util.cache_from_source -> importlib.util.cache_from_source

works in 3.6.9 at least and doc claims they were introduced in
3.2/3.4.


Geesh what a pain. Good thing there was a dockerfile with 3.12rc2 I
could use to debug this.