build(docker): disable pip_sdist mode
in github ci:

  COPY README.txt dist/* install/dist/

crashes the docker build if the dist directory is missing. This
prevents anchore/grype from scanning the image.

This is a new issue as it doesn't happen on my system.  The README.txt
file is there explicitly to prevent COPY from failing if nothing
matches the wildcard. This used to work, but ....

So I am disabling building from local sdist.

This only affects maintainers. Building from sdist is done to generate
a docker that matches a source distribution that has yet to be uploaded
to pypi.

Other use cases can build using the default of local.

I haven't found a way to make a dist subdir in the docker build
context directory. I can't even figure out how to identify the context
directory name inside the build container.

Also updated installation docs.