fix: exit quickly on keyboard interrupt
When exiting roundup_server (issue was seen in use of roundup_demo)
using ^C, the keyboard interrupt message is generated but the
application didn't exit until the 60 second socket timeout is
complete.

This seems to be more of an issue with Windows.

With this change the socket is explicitly shut down telling the client
what's happening. Then the file descriptor is closed.