clarify doc on dispatcher_email config setting.
An issue was brought up on the mailing list.

https://sourceforge.net/p/roundup/mailman/message/43383465/

The description of dispatcher_email sounds like it should be sent
email on issue creation. That's not it's role. Try to make it's role
more obvious.

Fix config.ini and reference.txt description.

Add the newissuecopy.py detector to send email on the creation of an
issue