doc: add comment on method to set savepoint_limit dynamically
savepoint_limit isn't a real config option as in config.ini. I
generate one on the fly for use in the backend during import.

While this is arguably a valid config option, it is also something
that is niche (import a tracker to postgres) and adds clutter to
config.ini.