fix: make task index not crash.
It looks like the task.index.html page was a copy of bug.index.html.
However some fields are missing from a task that is in a bug.  This
causes a crash when looking at all tasks if you don't explicitly
select the fields to be displayed.

The task links from the sidebar do explicitly select @columns fields,
so the crash is not normally seen.