fix: windows install using pip mislocates share directory
The setup code that tries to make the share install path absolute
prependeds something like:

  c:\program files\python_venv

to the paths. The equivalent on linux is recognized as an absolute
path. On windows this is treated oddly. This resulted in
the share files being placed in:

  c:\program files\python_venv\Lib\site-packages\program files\python_venv\share

Roundup was unable to find the files there. On windows (where the
platform starts with 'win') don't make the path absolute. This puts
share in:

  c:\program files\python_venv\Lib\share

and Roundup finds them.

The translations and templates are found by the roundup-server.
The docs are also installed under the share directory. The man pages
are not installed as windows doesn't have groff to format the source
documents.

This is the second fix from issues getting Roundup running on windows
discussed on mailing list by Simon Eigeldinger.

Thread starts with:

 https://sourceforge.net/p/roundup/mailman/message/41557096/

subject:  Installing Roundup on Windows 2023-10-05.