fix: roundup-demo, interactive mode would nuke an existing tracker.
Tonu Mikk demonstrated a bug in roundup-demo. When invoked without a
tracker home on the command line, it would nuke an existing directory
specified interactively.

There is still a minor bug. If an invalid home directory is specified,
roundup-demo will prompt for the template to use before reporting:

  Error: Refusing to nuke non-tracker directory:

but this doesn't cause data loss so not worth fixing at this time.