doc: add FTS5 requirement for sqlite backend.
Document the requirement for FTS5 if you are using the sqlite backend.

Tonu Mikk reported issues with using a SQLite that was missing FTS5
support (RedHat 7 sqlite with python 3.6). Even though FTS5 has been
in the sqlite amalgamation since 2015-10-14 (v3.9.0) RedHat chose to
not build their package with that. This leads to a traceback when
initializing the database.