test: actually test output from running demo server code
I was invoking demo mode, but not testing the output to verify it was
making it to the point where it would start the server.

Check that an expected database file was created to verify that the db
setting was honored.

Also test TEMPLATE-INFO.txt to verify that the correct template was
being instantiated.

Tested under 3.13 as wel using docker command line (wrapped):

   docker run -it -u 1000 --rm -v $PWD:/usr/src/myapp -w
   /usr/src/myapp  python:3.13.0a1-alpine3.18 sh -c 'export
   HOME=/tmp/home; mkdir $HOME; python -m pip install pytest
   pytest-env requests jinja2; python -m pytest -v test/test_demo.py'

If demo mode tries to start as root, it exits with an error, so it
must be run with -u uid.