doc: replace URL action example with form
The example to provide a link to take or resolve an item was broken.
Roundup has required a POST not GET for changes to the database for
a while. This example was never fixed.

Replaced a link using GET with a POSTed form. This changes the
activating element from a link to a button. The button could be
restyled as a link using CSS I think.

The original example was also not canonical. It used :action rather
than the canonical @action.