doc: document lib, fix formating/example
Add lib as an optional directory in a template.

Make config_ini.ini example an actual code block using the file from
the jinja2 template.

Format the TEMPLATE-INFO.txt example as plain text (it was
highlighting words like 'in' and 'with'). Also move the whole example
under the bullet item, it was outside of the list. Reformat the input
text to be less than 80 characters.