doc: trying to get search working.
Searching for www works beter now:

 * no more javscript errors in console
 * search context blocks are displayed

Had to add role="main" as that is used by the javascript to id
the body of the page to extract text snippets.

Had to set script_files to load all the needed javascript in order.

Had to set data-url_root on a id'ed tag for javacript to work.

On doc/_templates/layout.html replaced div with main tag and set id
and duplicate role="main" for use by js querySelector.