doc: make top of table columns align
Added:

  .. table::
    :class: valign-top

in multiple places. It adds the valign-top class to each table to make
the left hand column (label/index) align with the top of the
description in the right hand column.

This prevents the label from being 1/2 way down the screen when there is
a large description that starts at the top of the screen.

This needs to be done throughout the docs. rest.txt already has it done.