fix: cgitb crashes if there is no line number set.
I am not sure why:

  context.position[0]

is None when an exception happens while processing templates. This
causes a second exception because of a print statement matching %d to
None.  The second exception blows away any chance at getting useful
info.  So fix it by passing line number -1. Hopefully people will
recognize that as bogus and interpret it as no line number was
available.

Fixing the underlying cause (in ZopeTAL???) would be better though.