test: add pyproject.toml: set pytest and ruff project defaults
Pytest 6 is required to use this. This means it works only for python3.

It includes a dummy setting for redis password assuming redis is
running at the default port. This requires pytest-env be installed.

At some point I expect to add project and build-system tables to allow
building using the build package. The version and long_description
specifiers in the project table will be dynamic since the version is
from roundup/__init__.py and description from doc/announcement.txt.

One open issue is how to support the equivalent of:

  python setup.py build_doc

to run sphinx over the docs and include them in the share directory
bundled with the distributions.

Modified test_redis_session.py to better handle empty password.