test: add hypothesis
Allow python3 tests to use the new hypothesis testing.

I am targeting python 3 for additional testing of existing code.
Don't spend the additional time on Travis CI (where python2 is tested)
where I have to ask for more time on the platform.

New code should be testable on both platforms. If we have a need to
test with hypothesis on python 2 for new code, we can revisit this.